Fixed pen tablet being unresponsible on the color select widget.
The behaviour was that when you click, nothing happened, and only after dragging a certain distance does something happen. Turns out this is because of Windows "pen drag threshold" that forces you to move a certain distance before it registers a click. To fix this, I just reroute all pen movements as mouse movements, so ImGUI sees the mouse movement and just treats the pen as a mouse.