Fixed grid like artifacts when drawing
It seems that writing fragment shaders in a way where we sample from the texture at the same time as we write to it causes race conditions, which causes strange grid like patterns. The fix was using OpenGL's blend modes to handle this instead, thus removing the need to sample the texture as we write to it.